+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.5.7] - 2025-04-08
+### Added
+- Button in calling dashboard to go to the first patiënt who has not yet been called.
+- Warning to first save results in the calling dashboard, to prevent that the call_status remains 'wordt gebeld'. 
+
+### Changed
+- Mute_list query to exclude patients who are called today, so that these patients are still visible in the calling dashboard. 
+
+### Removed
+- Selectbox for user of calling dashboard to select the call_status. 
+
 
 ## [1.5.6] - 2025-03-31